Move dependencies to `dependencies.py`.
